Remove webactivator dependency and switch to initialize pipeline
