+++ Source/Synthesis/Synthesis.csproj
@@ -40,10 +40,6 @@
     </Reference>
     <Reference Include="Microsoft.Build" />
     <Reference Include="Microsoft.Build.Framework" />
-    <Reference Include="Microsoft.Web.Infrastructure, Version=1.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
-      <Private>True</Private>
-      <HintPath>..\..\packages\Microsoft.Web.Infrastructure.1.0.0.0\lib\net40\Microsoft.Web.Infrastructure.dll</HintPath>
-    </Reference>
     <Reference Include="Sitecore.ContentSearch">
       <HintPath>..\..\Dependencies\Libraries\Sitecore\Sitecore.ContentSearch.dll</HintPath>
     </Reference>
@@ -65,9 +61,6 @@
     <Reference Include="System.Core" />
     <Reference Include="System.Web" />
     <Reference Include="System.Xml" />
-    <Reference Include="WebActivatorEx">
-      <HintPath>..\..\packages\WebActivatorEx.2.0.6\lib\net40\WebActivatorEx.dll</HintPath>
-    </Reference>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="..\SharedAssemblyInfo.cs">
@@ -140,11 +133,12 @@
     <Compile Include="ItemExtensions.cs" />
     <Compile Include="Initializers\ITemplateInitializer.cs" />
     <Compile Include="MissingTemplateFieldException.cs" />
+    <Compile Include="Pipelines\Initialize\CheckModelSynchronization.cs" />
+    <Compile Include="Pipelines\Initialize\RegisterDefaultConfiguration.cs" />
     <Compile Include="SynthesisEditContext.cs" />
     <Compile Include="Synchronization\ModelTemplateReference.cs" />
     <Compile Include="Synchronization\SyncSource.cs" />
     <Compile Include="Templates\ITemplateFieldInfo.cs" />
-    <Compile Include="App_Start\SynthesisStartup.cs" />
     <Compile Include="Templates\ITemplateInfo.cs" />
     <Compile Include="Templates\ITemplateInputProvider.cs" />
     <Compile Include="Templates\ITemplateSignatureProvider.cs" />
