--- Source/Synthesis/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net45" />
-  <package id="WebActivatorEx" version="2.0.6" targetFramework="net45" />
-</packages>
